OPENAI/requirements.txt
twilio
flask
requests
openai==0.28
pyngrok
python-dotenv
